Remove webpack build in preparation for using elm reactor and elm make based build
