+++ copy-static.sh
@@ -1,0 +1,3 @@
+#!/bin/sh
+
+cp src/assets/* dist/
