--- package.json
@@ -1,22 +1,0 @@
-{
-  "devDependencies": {
-    "css-loader": "^0.28.11",
-    "elm-webpack-loader": "^4.5.0",
-    "file-loader": "^1.1.11",
-    "html-loader": "^0.5.5",
-    "html-webpack-plugin": "^3.2.0",
-    "node-sass": "^4.9.0",
-    "sass-loader": "^7.0.1",
-    "style-loader": "^0.21.0",
-    "url-loader": "^1.0.1",
-    "webpack": "^4.8.2",
-    "webpack-cli": "^2.1.3",
-    "webpack-serve": "^0.3.2"
-  },
-  "scripts": {
-    "build-prod": "webpack --config ./webpack.config.prod.js",
-    "build-dev": "webpack --config ./webpack.config.dev.js",
-    "build-elm": "webpack --config ./webpack.config.elm.js",
-    "start": "webpack-serve ./webpack.config.dev.js"
-  }
-}
