--- src/assets/index-elm.html
@@ -1,20 +1,0 @@
-<!DOCTYPE html>
-<html lang="en">
-    <head>
-        <meta charset="utf-8">
-        <meta name="viewport" content="width=device-width, initial-scale=1, shrink-to-fit=no">
-        <meta name="theme-color" content="#000000">
-        <title>ChoicEats</title>
-        <link rel="//fonts.googleapis.com/css?family=Titillium+Web:700|Source+Serif+Pro:400,700|Merriweather+Sans:400,700|Source+Sans+Pro:400,300,600,700,300italic,400italic,600italic,700italic" rel="stylesheet" type="text/css">
-        <link rel="stylesheet" href="https://cdnjs.cloudflare.com/ajax/libs/semantic-ui/2.3.0/semantic.min.css"><!---->
-
-    </head>
-    <body id="page-body">
-        <noscript>
-            Enable Javascript to run this app.
-        </noscript>
-        <script src="elm-dev-bundle.js" type="text/javascript"></script>
-        <script src="elm-dev-bootstrap.js" type="text/javascript"></script>
-    </body>
-</html>
-
