+++ src/assets/manifest.json
@@ -1,6 +1,6 @@
 {
-  "short_name": "React App",
-  "name": "Create React App Sample",
+  "short_name": "ChoicEats",
+  "name": "Delightful Recipes",
   "icons": [
     {
       "src": "favicon.ico",
