--- src/elm-package.json
@@ -1,14 +1,0 @@
-{
-    "version": "4.0.0",
-    "summary": "Example Project",
-    "repository": "https://github.com/user/project.git",
-    "license": "BSD-3-Clause",
-    "source-directories": [
-        "src"
-    ],
-    "exposed-modules": [],
-    "dependencies": {
-        "elm-lang/core": "5.0.0 <= v < 6.0.0"
-    },
-    "elm-version": "0.18.0 <= v < 0.19.0"
-}
