--- webpack.config.common.js
@@ -1,59 +1,0 @@
-const path = require('path');
-const HtmlWebPackPlugin = require('html-webpack-plugin');
-
-module.exports = {
-    entry: {
-        app: [
-            './src/assets/index.js'
-        ]
-    },
-
-    output: {
-        path: path.resolve(__dirname + '/dist'),
-        filename: '[name].js',
-    },
-
-    module: {
-        rules: [
-            {
-                test: /\.(css|scss)$/,
-                use: [
-                    'style-loader',
-                    'css-loader',
-                    'sass-loader',
-                ]
-            },
-            {
-                test: /\.html$/,
-                use: [
-                    {
-                        loader: 'html-loader',
-                        options: { minimize: true }
-                    }
-                ]
-            },
-            {
-                test: /\.elm$/,
-                exclude: [/elm-stuff/, /node_modules/],
-                loader: 'elm-webpack-loader?verbose=true&warn=true',
-            },
-            {
-                test: /\.woff(2)?(\?v=[0-9]\.[0-9])?$/,
-                loader: 'url-loader?limit=10000^mimetype=application/font-woff',
-            },
-            {
-                test: /\.(ttf|eot|svg|ico)(\?v=[0-9]\.[0-9]\.[0-9])?$/,
-                loader: 'file-loader',
-            },
-        ],
-
-        noParse: /\.elm$/,
-    },
-
-    plugins: [
-        new HtmlWebPackPlugin({
-            template: './src/assets/index.html',
-            filename: './index.html',
-        }),
-    ],
-}
