--- windows-dev-server.bat
@@ -1,2 +1,0 @@
-REM Currently broken
-elm-live --output=public/elm.js src/Main.elm --dir=public/ --before-build=cls --pushstate --open --debug
